# docs/swagger.yaml
definitions:
  controllers.allCategoryResponse:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  controllers.authForm:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  controllers.authResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  controllers.createCategoryForm:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  controllers.createUserForm:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  controllers.pagingResult:
    properties:
      count:
        type: integer
      limit:
        type: integer
      nextPage:
        type: integer
      page:
        type: integer
      prevPage:
        type: integer
      totalPage:
        type: integer
    type: object
  controllers.productResponse:
    properties:
      category:
        properties:
          id:
            type: integer
          name:
            type: string
        type: object
      categoryId:
        type: integer
      desc:
        type: string
      id:
        type: integer
      image:
        type: string
      name:
        type: string
      price:
        type: integer
    required:
    - categoryId
    type: object
  controllers.productsPaging:
    properties:
      items:
        items:
          $ref: '#/definitions/controllers.productResponse'
        type: array
      paging:
        $ref: '#/definitions/controllers.pagingResult'
    type: object
  controllers.updateCategoryForm:
    properties:
      name:
        type: string
    type: object
  controllers.updateUserForm:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
    type: object
  controllers.userResponse:
    properties:
      avatar:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      role:
        type: string
    type: object
  middleware.login:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
info:
  contact: {}
  title: Swagger GO-API-PRODUCT API
  version: "1.0"
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: login by form user
      parameters:
      - description: login
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/middleware.login'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: login
      tags:
      - auth
  /api/v1/auth/profile:
    get:
      consumes:
      - application/json
      description: get by form user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.userResponse'
      security:
      - BearerAuth: []
      summary: get an user profile
      tags:
      - auth
    put:
      consumes:
      - multipart/form-data
      description: add by form user
      parameters:
      - description: name
        in: formData
        name: name
        required: true
        type: string
      - description: email
        in: formData
        name: email
        required: true
        type: string
      - description: avatar
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.userResponse'
      security:
      - BearerAuth: []
      summary: update an user
      tags:
      - auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: add by form user
      parameters:
      - description: authForm
        in: body
        name: authForm
        required: true
        schema:
          $ref: '#/definitions/controllers.authForm'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controllers.authResponse'
      summary: add an user
      tags:
      - auth
  /api/v1/categories:
    get:
      consumes:
      - application/json
      description: get by form categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/controllers.allCategoryResponse'
            type: array
      summary: Show an categories
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: add by form category
      parameters:
      - description: name
        in: body
        name: name
        required: true
        schema:
          $ref: '#/definitions/controllers.createCategoryForm'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controllers.allCategoryResponse'
      security:
      - BearerAuth: []
      summary: add an category
      tags:
      - categories
  /api/v1/categories/{id}:
    delete:
      consumes:
      - application/json
      description: delete by json category
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "404":
          description: '{"error": "not found"}'
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Bad Request
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: delete an category
      tags:
      - categories
    get:
      consumes:
      - application/json
      description: get by form category
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.allCategoryResponse'
      summary: Show an category
      tags:
      - categories
    patch:
      consumes:
      - application/json
      description: update by form category
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: name
        in: body
        name: name
        required: true
        schema:
          $ref: '#/definitions/controllers.updateCategoryForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.allCategoryResponse'
      security:
      - BearerAuth: []
      summary: update an category
      tags:
      - categories
  /api/v1/products:
    get:
      consumes:
      - application/json
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.productsPaging'
      summary: Show an products
      tags:
      - products
    post:
      consumes:
      - multipart/form-data
      description: add by form product
      parameters:
      - description: name
        in: formData
        name: name
        required: true
        type: string
      - description: desc
        in: formData
        name: desc
        required: true
        type: string
      - description: price
        in: formData
        name: price
        required: true
        type: integer
      - description: image
        in: formData
        name: image
        required: true
        type: file
      - description: categoryId
        in: formData
        name: categoryId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controllers.productResponse'
      security:
      - BearerAuth: []
      summary: add an product
      tags:
      - products
  /api/v1/products/{id}:
    delete:
      consumes:
      - application/json
      description: delete by json product
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.productResponse'
        "404":
          description: '{"error": "not found"}'
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Bad Request
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: delete an product
      tags:
      - products
    get:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.productResponse'
      summary: FindOne - /:id
      tags:
      - products
    patch:
      consumes:
      - multipart/form-data
      description: update by form product
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: name
        in: formData
        name: name
        type: string
      - description: desc
        in: formData
        name: desc
        type: string
      - description: price
        in: formData
        name: price
        type: integer
      - description: image
        in: formData
        name: image
        type: file
      - description: categoryId
        in: formData
        name: categoryId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.productResponse'
      security:
      - BearerAuth: []
      summary: update an products
      tags:
      - products
  /api/v1/users:
    get:
      consumes:
      - application/json
      description: get by form users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/controllers.userResponse'
            type: array
      security:
      - BearerAuth: []
      summary: get an users
      tags:
      - user
    post:
      consumes:
      - application/json
      description: add by form user
      parameters:
      - description: create-user-form
        in: body
        name: createUserForm
        required: true
        schema:
          $ref: '#/definitions/controllers.createUserForm'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controllers.userResponse'
      security:
      - BearerAuth: []
      summary: add an user
      tags:
      - user
  /api/v1/users/{id}:
    delete:
      consumes:
      - application/json
      description: delete by form user
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: ""
      security:
      - BearerAuth: []
      summary: delete an user
      tags:
      - user
    get:
      consumes:
      - application/json
      description: get by form user
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.userResponse'
      security:
      - BearerAuth: []
      summary: get an user
      tags:
      - user
    patch:
      consumes:
      - application/json
      description: update by form user
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: update-user-form
        in: body
        name: updateUserForm
        schema:
          $ref: '#/definitions/controllers.updateUserForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.userResponse'
      security:
      - BearerAuth: []
      summary: update an user
      tags:
      - user
  /api/v1/users/{id}/demote:
    patch:
      consumes:
      - application/json
      description: update by form user
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.userResponse'
      security:
      - BearerAuth: []
      summary: update an user
      tags:
      - user
  /api/v1/users/{id}/promote:
    patch:
      consumes:
      - application/json
      description: update by form user
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.userResponse'
      security:
      - BearerAuth: []
      summary: update an user
      tags:
      - user
  /healthz:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: ""
  /x:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"